Create component OpenUrlButton
