+++ packages/mobile/src/screens/Login/Login.js
@@ -1,25 +1,7 @@
-import {View, TextInput, ImageBackground, Pressable, Text, Linking, Alert} from "react-native";
+import {View, TextInput, ImageBackground, Pressable, Text} from "react-native";
 import styles from "./LoginStyle";
 import {useFonts} from "expo-font";
-import {useCallback} from "react";
-
-const OpenURLButton = ({ url, children }) => {
-    const handlePress = useCallback(async () => {
-        // Checking if the link is supported for links with custom URL scheme.
-        const supported = await Linking.canOpenURL(url);
-
-        if (supported) {
-            // Opening the link with some app, if the URL scheme is "http" the web link should be opened
-            // by some browser in the mobile
-            await Linking.openURL(url);
-        } else {
-            Alert.alert(`Impossible d'ouvrir le lien : ${url}`);
-        }
-    }, [url]);
-
-    return <Text onPress={handlePress} style={styles.forgetPasswordText}>{children}</Text>;
-};
-
+import OpenURLButton from "../../components/OpenUrlButton";
 
 export default function Login() {
     const [loaded] = useFonts({
@@ -50,7 +32,7 @@ export default function Login() {
                     />
 
                     <OpenURLButton url={'http://localhost:3000/forgot-password/'}>Mot de passe oublié ?</OpenURLButton>
-                    <Pressable style={styles.button} onPress={() => alert('Marwane le gros suceur')}>
+                    <Pressable style={styles.button} onPress={() => alert('Test')}>
                         <Text style={styles.textButton}>Se connecter</Text>
                     </Pressable>
                 </View>
